refactor: 테스트 profile 수정
